Update demo to meteor 1.8.1 + package fixes
