+++ demo/.meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.3.2.4
+METEOR@1.8.1
