+++ demo/.meteor/versions
@@ -1,79 +1,93 @@
-accounts-base@1.2.7
-accounts-ui@1.1.9
-accounts-ui-unstyled@1.1.12
-allow-deny@1.0.4
-autoupdate@1.2.9
-babel-compiler@6.6.4
-babel-runtime@0.1.8
-base64@1.0.8
-binary-heap@1.0.8
-blaze@2.1.7
-blaze-tools@1.0.8
-boilerplate-generator@1.0.8
-caching-compiler@1.0.4
-caching-html-compiler@1.0.6
-callback-hook@1.0.8
-check@1.2.1
-ddp@1.2.5
-ddp-client@1.2.7
-ddp-common@1.2.5
-ddp-rate-limiter@1.0.4
-ddp-server@1.2.6
+accounts-base@1.4.4
+accounts-ui@1.3.1
+accounts-ui-unstyled@1.4.2
+allow-deny@1.1.0
+autoupdate@1.6.0
+babel-compiler@7.3.4
+babel-runtime@1.3.0
+base64@1.0.12
+binary-heap@1.0.11
+blaze@2.3.3
+blaze-html-templates@1.1.2
+blaze-tools@1.0.10
+boilerplate-generator@1.6.0
+caching-compiler@1.2.1
+caching-html-compiler@1.1.3
+callback-hook@1.1.0
+check@1.3.1
+coffeescript@1.0.17
+ddp@1.4.0
+ddp-client@2.3.3
+ddp-common@1.4.0
+ddp-rate-limiter@1.0.7
+ddp-server@2.3.0
 deps@1.0.12
-diff-sequence@1.0.5
-ecmascript@0.4.3
-ecmascript-runtime@0.2.10
-ejson@1.0.11
-fastclick@1.0.11
-geojson-utils@1.0.8
-html-tools@1.0.9
-htmljs@1.0.9
-http@1.1.5
-id-map@1.0.7
-jquery@1.11.8
-launch-screen@1.0.11
-less@2.6.0
+diff-sequence@1.1.1
+dynamic-import@0.5.1
+ecmascript@0.12.7
+ecmascript-runtime@0.7.0
+ecmascript-runtime-client@0.8.0
+ecmascript-runtime-server@0.7.1
+ejson@1.1.0
+es5-shim@4.8.0
+fetch@0.1.1
+geojson-utils@1.0.10
+hot-code-push@1.0.4
+html-tools@1.0.11
+htmljs@1.0.11
+http@1.4.2
+id-map@1.1.0
+inter-process-messaging@0.1.0
+jquery@1.11.11
+launch-screen@1.1.1
+less@2.8.0
 livedata@1.0.18
-localstorage@1.0.9
-logging@1.0.12
-meteor@1.1.14
-meteor-platform@1.2.6
-minifier-css@1.1.11
-minifier-js@1.1.11
-minimongo@1.0.16
+localstorage@1.2.0
+logging@1.1.20
+meteor@1.9.3
+meteor-base@1.4.0
+minifier-css@1.4.2
+minifier-js@2.4.1
+minimongo@1.4.5
 mizzao:accounts-testing@0.1.0
-mizzao:timesync@0.4.0
-mizzao:user-status@0.6.6
-mobile-status-bar@1.0.12
-modules@0.6.1
-modules-runtime@0.6.3
-mongo@1.1.7
-mongo-id@1.0.4
-mrt:moment@2.8.1
-npm-mongo@1.4.43
-observe-sequence@1.0.11
-ordered-dict@1.0.7
-promise@0.6.7
-random@1.0.9
-rate-limit@1.0.4
-reactive-dict@1.1.7
-reactive-var@1.0.9
-reload@1.1.8
-retry@1.0.7
-routepolicy@1.0.10
-service-configuration@1.0.9
-session@1.1.5
-spacebars@1.0.11
-spacebars-compiler@1.0.11
-standard-app-packages@1.0.9
-standard-minifier-css@1.0.6
-standard-minifier-js@1.0.6
-templating@1.1.9
-templating-tools@1.0.4
-tracker@1.0.13
+mizzao:timesync@0.5.1
+mizzao:user-status@0.7.0
+mobile-experience@1.0.5
+mobile-status-bar@1.0.14
+modern-browsers@0.1.4
+modules@0.13.0
+modules-runtime@0.10.3
+mongo@1.6.3
+mongo-decimal@0.1.1
+mongo-dev-server@1.1.0
+mongo-id@1.0.7
+npm-mongo@3.1.2
+observe-sequence@1.0.16
+ordered-dict@1.1.0
+promise@0.11.2
+random@1.1.0
+rate-limit@1.0.9
+reactive-dict@1.3.0
+reactive-var@1.0.11
+reload@1.3.0
+retry@1.1.0
+routepolicy@1.1.0
+service-configuration@1.0.11
+session@1.2.0
+shell-server@0.4.0
+socket-stream-client@0.2.2
+spacebars@1.0.15
+spacebars-compiler@1.1.3
+standard-minifier-css@1.5.3
+standard-minifier-js@2.4.1
+templating@1.3.2
+templating-compiler@1.3.3
+templating-runtime@1.3.2
+templating-tools@1.1.2
+tracker@1.2.0
 twbs:bootstrap@3.3.6
-ui@1.0.11
-underscore@1.0.8
-url@1.0.9
-webapp@1.2.8
+ui@1.0.13
+underscore@1.0.10
+url@1.2.0
+webapp@1.7.4
 webapp-hashing@1.0.9
