+++ demo/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "user-status-demo",
+  "private": false,
+  "scripts": {
+    "lint": "./node_modules/.bin/eslint .",
+    "test": "meteor test --once --driver-package meteortesting:mocha",
+    "test-app": "TEST_WATCH=1 meteor test --full-app --driver-package meteortesting:mocha",
+    "visualize": "meteor --production --extra-packages bundle-visualizer"
+  },
+  "dependencies": {
+    "@babel/runtime": "^7.6.0",
+    "moment": "^2.24.0",
+    "underscore": "^1.9.1"
+  },
+  "devDependencies": {
+    "autoprefixer": "^9.6.1",
+    "babel-eslint": "^10.0.3"
+  },
+  "meteor": {
+    "mainModule": {
+      "client": "client/main.js",
+      "server": "server/main.js"
+    }
+  }
+}
