+++ demo/packages/meteor-user-status/monitor.js
@@ -1,0 +1,1 @@
+../../../monitor.js
