+++ demo/packages/meteor-user-status/package.js
@@ -1,0 +1,1 @@
+../../../package.js
