+++ demo/packages/meteor-user-status/status.js
@@ -1,0 +1,1 @@
+../../../status.js
