--- demo/smart.json
@@ -1,10 +1,0 @@
-{
-  "packages": {
-    "mizzao:timesync": {
-      "path": "../../mizzao:timesync"
-    },
-    "mizzao:user-status": {
-      "path": ".."
-    }
-  }
-}
