--- demo/smart.lock
@@ -1,21 +1,0 @@
-{
-  "meteor": {},
-  "dependencies": {
-    "basePackages": {
-      "mizzao:timesync": {
-        "path": "../../mizzao:timesync"
-      },
-      "mizzao:user-status": {
-        "path": ".."
-      }
-    },
-    "packages": {
-      "mizzao:timesync": {
-        "path": "../../mizzao:timesync"
-      },
-      "mizzao:user-status": {
-        "path": ".."
-      }
-    }
-  }
-}
