+++ package.json
@@ -1,8 +1,14 @@
 {
   "name": "meteor-user-status",
+  "version": "0.7.0",
   "repository": {
-    "url": "https://github.com/Meteor-Community-Packages/meteor-user-status"
+    "type": "git",
+    "url": "git+https://github.com/Meteor-Community-Packages/meteor-user-status.git"
   },
+  "bugs": {
+    "url": "https://github.com/Meteor-Community-Packages/meteor-user-status/issues"
+  },
+  "homepage": "https://github.com/Meteor-Community-Packages/meteor-user-status#readme",
   "private": false,
   "scripts": {
     "lint": "./node_modules/.bin/eslint .",
@@ -17,7 +23,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "npm run lint && npm run test"
+      "pre-commit": "npm run test"
     }
   }
 }
